rag-service/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
sqlalchemy==2.1.4
databases==0.6.0
pydantic==1.10.12
python-multipart==0.0.6
sentence-transformers==2.2.2
tqdm==4.66.1
python-docx==0.8.11
pdfminer.six==20240713
requests==2.31.0
openai==0.29.2
aiofiles==23.1.0
chroma-db==0.5.9
*** End Patch
